Use dtype=float for many of the numpy.sum calls for older numpy versions where that wasn't the default (#845)

+++ galsim/noise.py
@@ -72,7 +72,7 @@ def addNoiseSNR(self, noise, snr, preserve_flux=False):
     @returns the variance of the noise that was applied to the image.
     """
     noise_var = noise.getVariance()
-    sumsq = np.sum(self.array**2)
+    sumsq = np.sum(self.array**2, dtype=float)
     if preserve_flux:
         new_noise_var = sumsq/snr/snr
         noise = noise.withVariance(new_noise_var)
